Do not use a pointer cursor for the button elements by default
Using an inline style for the cursor makes it impossible to set
different value when needed without using `!important`. This is
a case with, for example, disabled buttons which should have
`cursor: not-allowed` style.

CSS specification for the cursor[1] says that `pointer` value
should be used for indicating a link. In many cases, the button
selector have a different role than redirecting to another
location.

[1]: https://www.w3.org/TR/css-ui-3/#cursor